rename provider and provisioner packages
Using plural names to avoid collisions in existing code.
