--- provider/doc.go
@@ -1,3 +1,0 @@
-// Package provider contains the interface and primary types required to
-// implement a Terraform resource provider.
-package provider
